feat(info): count git commits in repo count and current branch
In former versions, `gitCommitCount` was mistekenly implemented to count
all commits in repo, not only on current branch.

This commit added `gitCommitCountRepo` and `gitCommitCountCurrentBranch`
to provide both commints count in repo and on current branch.

* `gitCommitCountRepo` acts like `gitCommitCount` in former versions,
  counting commits in repo, all branches.
* `gitCommitCountCurrentBranch` only counts commits on current branch
  which was intended.

This commit also marks `gitCommitCount` as deprecated because it may
mislead users, suggest to use the new added two variables on demand.